requirements.txt
numpy==1.17.0
torch==1.7.0
torchsummary==1.5.1
pandas==1.0.5
tqdm==4.47.0
prefetch_generator==1.0.1
opencv_python_headless==4.4.0.46
torchvision==0.8.1
matplotlib==3.2.2
keras==2.2.4
Pillow==8.0.1
scikit_learn==0.23.2
